Move fallback mapping to its own extension method
