+++ src/ForEvolve.ExceptionMapper.CommonHttpExceptionHandlers/Fallback/FallbackExceptionHandlerOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace ForEvolve.ExceptionMapper
+﻿namespace ForEvolve.ExceptionMapper.Handlers.Fallback
 {
     public class FallbackExceptionHandlerOptions
     {
